requirements.txt
pip==25.2
websocket-client==1.9.0
